refactor: ES 기반 상품 검색 모듈을 support 서비스로 분리
