--- catalog-service/src/main/java/com/node5/catalogservice/config/search/SearchIndexProperties.java
@@ -1,13 +1,0 @@
-package com.node5.catalogservice.config.search;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@ConfigurationProperties(prefix = "app.search.index")
-public class SearchIndexProperties {
-	private String product;
-}
