--- catalog-service/src/main/java/com/node5/catalogservice/search/application/dto/ProductAutocompleteCommand.java
@@ -1,4 +1,0 @@
-package com.node5.catalogservice.search.application.dto;
-
-public record ProductAutocompleteCommand(String keyword) {
-}
