--- catalog-service/src/main/java/com/node5/catalogservice/search/application/port/ProductSearchPort.java
@@ -1,17 +1,0 @@
-package com.node5.catalogservice.search.application.port;
-
-import java.util.List;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-import com.node5.catalogservice.search.application.dto.ProductAutocompleteCommand;
-import com.node5.catalogservice.search.application.dto.ProductSearchCommand;
-import com.node5.catalogservice.search.domain.ProductDocument;
-
-public interface ProductSearchPort {
-
-	Page<ProductDocument> search(ProductSearchCommand command, Pageable pageable);
-
-	List<String> autocomplete(ProductAutocompleteCommand command);
-}
