--- catalog-service/src/main/java/com/node5/catalogservice/search/application/query/QueryNormalizer.java
@@ -1,5 +1,0 @@
-package com.node5.catalogservice.search.application.query;
-
-public interface QueryNormalizer {
-	String normalize(String raw);
-}
