--- catalog-service/src/main/java/com/node5/catalogservice/search/infrastructure/ProductSearchRepository.java
@@ -1,8 +1,0 @@
-package com.node5.catalogservice.search.infrastructure;
-
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-
-import com.node5.catalogservice.search.domain.ProductDocument;
-
-public interface ProductSearchRepository extends ElasticsearchRepository<ProductDocument, String> {
-}
