--- catalog-service/src/main/java/com/node5/catalogservice/search/infrastructure/kafka/KafkaProductIndexEventConsumer.java
@@ -1,52 +1,0 @@
-package com.node5.catalogservice.search.infrastructure.kafka;
-
-import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.stereotype.Component;
-
-import com.node5.catalogservice.product.event.ProductIndexEvent;
-import com.node5.catalogservice.search.domain.ProductDocument;
-import com.node5.catalogservice.search.infrastructure.ProductSearchRepository;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class KafkaProductIndexEventConsumer {
-
-	private final ProductSearchRepository productSearchRepository;
-
-	@KafkaListener(
-		topics = "${app.kafka.topics.product-index}",
-		groupId = "${spring.kafka.consumer.group-id:catalog-service-search}"
-	)
-	public void consume(ProductIndexEvent event) {
-
-		String productId = event.productId().toString();
-
-		log.info("Kafka 상품 색인 이벤트 수신, productId={}, type={}", productId, event.type());
-
-		try {
-			ProductDocument document = new ProductDocument(
-				productId,
-				event.shopId().toString(),
-				event.name(),
-				event.nameAutocomplete(),
-				event.category(),
-				event.thumbnailKey(),
-				event.price(),
-				event.status(),
-				event.createdAt()
-			);
-
-			productSearchRepository.save(document);
-
-			log.info("ES 상품 색인 완료, productId={}", productId);
-
-		} catch (Exception e) {
-			log.error("ES 상품 색인 실패, productId={}", productId, e);
-			throw e;
-		}
-	}
-}
