+++ config/config/catalog-service.yaml
@@ -25,10 +25,6 @@ spring:
     producer:
       key-serializer: org.apache.kafka.common.serialization.StringSerializer
       value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
-    consumer:
-      group-id: catalog-service-search
-      key-deserializer: org.springframework.kafka.support.serializer.ErrorHandlingDeserializer
-      value-deserializer: org.springframework.kafka.support.serializer.ErrorHandlingDeserializer
       properties:
         spring.deserializer.key.delegate.class: org.apache.kafka.common.serialization.StringDeserializer
         spring.deserializer.value.delegate.class: org.springframework.kafka.support.serializer.JsonDeserializer
@@ -44,10 +40,6 @@ app:
       product-index: product-index
       product-embedding: ai-service.product-embedding.v1
       product-discontinued: catalog-service.product-discontinued.v1
-
-  search:
-    index:
-      product: products
 
   s3:
     bucket: ${S3_BUCKET_NAME}
