+++ config/config/support-service.yaml
@@ -86,3 +86,8 @@ api:
 
 server:
   port: 8087
+
+app:
+  search:
+    index:
+      product: products
