+++ support-service/build.gradle
@@ -34,4 +34,6 @@ dependencies {
     implementation 'org.springframework.ai:spring-ai-core:1.0.0-M6'
 
     implementation 'org.springframework.boot:spring-boot-starter-batch'
+
+    implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
 }
