+++ support-service/src/main/java/com/node5/supportservice/config/SearchIndexNameConfig.java
@@ -1,0 +1,15 @@
+package com.node5.supportservice.config;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableConfigurationProperties(SearchIndexProperties.class)
+public class SearchIndexNameConfig {
+
+	@Bean(name = "productIndexName")
+	public String productIndexName(SearchIndexProperties props) {
+		return props.getProduct();
+	}
+}
