+++ support-service/src/main/java/com/node5/supportservice/config/SearchIndexProperties.java
@@ -1,0 +1,13 @@
+package com.node5.supportservice.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@ConfigurationProperties(prefix = "app.search.index")
+public class SearchIndexProperties {
+	private String product;
+}
