+++ support-service/src/main/java/com/node5/supportservice/search/application/dto/ProductAutocompleteCommand.java
@@ -1,0 +1,4 @@
+package com.node5.supportservice.search.application.dto;
+
+public record ProductAutocompleteCommand(String keyword) {
+}
