+++ support-service/src/main/java/com/node5/supportservice/search/application/query/QueryNormalizer.java
@@ -1,0 +1,5 @@
+package com.node5.supportservice.search.application.query;
+
+public interface QueryNormalizer {
+	String normalize(String raw);
+}
